# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: 邻里车位 API
  description: 阶段 2 免费车位共享 MVP 接口
  version: 0.2.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/owner/spaces/{id}:
    put:
      tags:
      - owner-space-controller
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpaceRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ParkingSpace'
  /api/v1/admin/communities/{id}:
    put:
      tags:
      - admin-controller
      operationId: updateCommunity
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Community'
...
  /api/v1/owner/spaces:
    get:
      tags:
      - owner-space-controller
      operationId: listSpaces
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkingSpace'
    post:
      tags:
      - owner-space-controller
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpaceRequest'
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ParkingSpace'
  /api/v1/owner/spaces/{spaceId}/slots:
    get:
      tags:
      - owner-space-controller
      operationId: listSlots
      parameters:
      - name: spaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailabilitySlot'
    post:
      tags:
      - owner-space-controller
      operationId: publishSlot
      parameters:
      - name: spaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlotRequest'
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AvailabilitySlot'
  /api/v1/bookings:
    get:
      tags:
      - booking-controller
      operationId: list_2
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingView'
    post:
      tags:
      - booking-controller
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookingView'
...
  /api/v1/auth/dev-login:
    post:
      tags:
      - auth-controller
      summary: 使用本地演示身份登录
      operationId: devLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevLoginRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoginResponse'
      security: []
  /api/v1/admin/spaces/{id}/reject:
    post:
      tags:
      - admin-controller
      operationId: reject
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewSpaceRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ParkingSpace'
  /api/v1/admin/spaces/{id}/approve:
    post:
      tags:
      - admin-controller
      operationId: approve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewSpaceRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ParkingSpace'
  /api/v1/admin/complaints/{id}/resolve:
    post:
      tags:
      - admin-controller
      operationId: resolveComplaint
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplaintResolveRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Complaint'
  /api/v1/admin/communities:
    get:
      tags:
      - admin-controller
      operationId: listCommunities
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Community'
    post:
      tags:
      - admin-controller
      operationId: createCommunity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityRequest'
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Community'
  /api/v1/spaces/{id}:
    get:
      tags:
      - space-controller
      operationId: detail
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: slotId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SearchResult'
  /api/v1/spaces/search:
    get:
      tags:
      - space-controller
      operationId: search
      parameters:
      - name: communityId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: startAt
        in: query
        required: true
        schema:
          type: string
          format: date-time
      - name: endAt
        in: query
        required: true
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
...
  /api/v1/owner/spaces/{spaceId}/slots/{slotId}:
    delete:
      tags:
      - owner-space-controller
      operationId: cancelSlot
      parameters:
      - name: spaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: slotId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
components:
  schemas:
...
    SpaceRequest:
      required:
      - accessInstructions
      - communityId
      - spaceCode
      - title
      - vehicleLimit
      type: object
      properties:
        communityId:
          type: integer
          format: int64
        spaceCode:
          maxLength: 64
          minLength: 0
          type: string
        title:
          maxLength: 128
          minLength: 0
          type: string
        accessInstructions:
          maxLength: 1000
          minLength: 0
          type: string
        vehicleLimit:
          maxLength: 64
          minLength: 0
          type: string
    ParkingSpace:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
        ownerId:
          type: integer
          format: int64
        communityId:
          type: integer
          format: int64
        spaceCode:
          type: string
        title:
          type: string
        accessInstructions:
          type: string
        vehicleLimit:
          type: string
        status:
          type: string
          enum:
          - PENDING_REVIEW
          - APPROVED
          - REJECTED
          - DISABLED
        reviewNote:
          type: string
    CommunityRequest:
      required:
      - address
      - latitude
      - longitude
      - name
      type: object
      properties:
        name:
          maxLength: 128
          minLength: 0
          type: string
        address:
          maxLength: 255
          minLength: 0
          type: string
        latitude:
          maximum: 90.0
          exclusiveMaximum: false
          minimum: -90.0
          exclusiveMinimum: false
          type: number
          format: double
        longitude:
          maximum: 180.0
          exclusiveMaximum: false
          minimum: -180.0
          exclusiveMinimum: false
          type: number
          format: double
        active:
          type: boolean
    Community:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        active:
          type: boolean
    SlotRequest:
      required:
      - endAt
      - startAt
      type: object
      properties:
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
    AvailabilitySlot:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
        spaceId:
          type: integer
          format: int64
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - PUBLISHED
          - CANCELLED
          - EXPIRED
        freeOfCharge:
          type: boolean
    CreateBookingRequest:
      required:
      - endAt
      - slotId
      - startAt
      - vehicleId
      type: object
      properties:
        slotId:
          type: integer
          format: int64
        vehicleId:
          type: integer
          format: int64
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
    BookingView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bookingNo:
          type: string
        status:
          type: string
          enum:
          - CONFIRMED
          - IN_USE
          - COMPLETED
          - CANCELLED
          - NO_SHOW
          - DISPUTED
        spaceId:
          type: integer
          format: int64
        spaceTitle:
          type: string
        spaceCode:
          type: string
        accessInstructions:
          type: string
        vehicleId:
          type: integer
          format: int64
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        cancelReason:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Complaint:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
        bookingId:
          type: integer
          format: int64
        complainantId:
          type: integer
          format: int64
        content:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - PROCESSING
          - RESOLVED
          - REJECTED
        resolutionNote:
          type: string
...
    DevLoginRequest:
      required:
      - userId
      type: object
      properties:
        userId:
          type: integer
          format: int64
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/UserView'
    UserView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nickname:
          type: string
        phoneMasked:
          type: string
        avatarUrl:
          type: string
        roles:
          uniqueItems: true
          type: array
          items:
            type: string
            enum:
            - USER
            - OWNER
            - PROPERTY_ADMIN
            - PLATFORM_ADMIN
    ReviewSpaceRequest:
      required:
      - note
      type: object
      properties:
        note:
          maxLength: 500
          minLength: 0
          type: string
    ComplaintResolveRequest:
      required:
      - note
      type: object
      properties:
        resolved:
          type: boolean
        note:
          maxLength: 1000
          minLength: 0
          type: string
    SearchResult:
      type: object
      properties:
        spaceId:
          type: integer
          format: int64
        slotId:
          type: integer
          format: int64
        title:
          type: string
        communityId:
          type: integer
          format: int64
        communityName:
          type: string
        communityAddress:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        maskedSpaceCode:
          type: string
        vehicleLimit:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        freeOfCharge:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT